Move claims to a public package
